requirements.txt
apache-airflow==2.8.0